Jp/save current (#20)
* feat:save file

* fix: format and xml write (unfinished)

* feat: clean script

* fix: rapidxml include errors

* fix: save to file (working)

* fix: cartesian2spherical (unfinished)

* feat: calculate rot unfinished

* feat: calculate rot unfinished

* feat: logs

* chore: formatting

+++ engine/src/parse.cpp
@@ -31,7 +31,7 @@ Configuration parseConfig(std::string filename) {
   char* width = root->first_node("window")->first_attribute("width")->value();
   char* height = root->first_node("window")->first_attribute("height")->value();
 
-  Window window_info = Window(std::stof(width), std::stof(height));
+  Window window_info = Window(std::stoi(width), std::stoi(height));
 
   // camera information
   rapidxml::xml_node<>* camera = root->first_node("camera");
